requirements_rpi.txt
tk~=0.1.0
setuptools~=51.3.3
Pillow~=8.2.0
Rpi.GPIO~=0.7.0
